Split type file into two files for the each route
