+++ server/routes/hicgenome.ts
@@ -1,4 +1,5 @@
-import { HicGenomeRequest, HicGenomeResponse, Item } from '#shared/types/routes/hic.ts'
+import { HicGenomeRequest, HicGenomeResponse } from '#shared/types/routes/hicgenome.ts'
+import { Item } from '#shared/types/routes/hicdata.ts'
 import { fileurl } from '#src/utils.js'
 import { spawn } from 'child_process'
 import readline from 'readline'
