+++ server/shared/types/routes/hicdata.ts
@@ -1,8 +1,6 @@
 import { FileORURL } from '../fileOrUrl.ts'
 
-/******** Types are for the hicdata and hicgenome endpoints ********/
-
-type BaseHicRequest = FileORURL & {
+export type BaseHicRequest = FileORURL & {
 	/** Value relates to the 1st parameter of straw tool, which accepts 'observed', 'expected', 'oe', 'norm', and 'distance' */
 	matrixType: 'observed' | 'expected' | 'oe' | 'log(oe)'
 	/** Either a base pair or fragment resolution calculated from the array*/
@@ -38,23 +36,4 @@ export type HicdataResponse = {
 	items: Item[]
 }
 
-export type HicGenomeRequest = BaseHicRequest & {
-	/** Entire chromosome list read from the file (see hicstate) */
-	chrlst: string[]
-	/** window location */
-	embedder: string
-	/** whether or not the file contains 'chr' for the chromosomes */
-	nochr: boolean
-}
 
-export type HicGenomeResponse = {
-	data: {
-		/** First chromosome */
-		lead: string
-		/** Second chromosome */
-		follow: string
-		items: Item[]
-	}[]
-	/** Error message to display on the client, if applicable */
-	error?: string
-}
